Move functions into util
And also refactored some of other components and pages.
