+++ packages/pwa/app/utils/responsive-image.js
@@ -5,83 +5,171 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {getAppOrigin} from 'pwa-kit-react-sdk/utils/url'
 import React from 'react'
 import {Image} from '@chakra-ui/react'
+import theme from '@chakra-ui/theme'
 
 /**
  * @param {Object} props
- * @param {string} props.src - The image's url
- * @param {Object} [props.sizes] - Defines how the image responds across Chakra breakpoints
- * @param {number[]} [props.widths] - Will be used to generate the image's srcset
- * @param {Object} [props.transformations] - How to transform the image's url {@link https://documentation.b2c.commercecloud.salesforce.com/DOC1/topic/com.demandware.dochelp/content/b2c_commerce/topics/image_management/b2c_creating_image_transformation_urls.html}
- * @return {Object} - Props for an image component
- *
- * @example
- * <Img
- *     {...getImageProps({
- *         src,
- *         sizes: {base: 'calc(100vw / 2)', md: '500px'},
- *         widths: [500, 1000, 1500],
- *         transformations: {q: 60}
- *     })}
- * />
+ * @param {string} props.src
+ * @param {number[]} props.vwSizes
+ * @param {(string[]|Object|string)} props.sizes
+ * @param {(number[]|string)} props.srcSet
+ * @return {Object} props for Chakra image component
  */
-export const getImageProps = ({src, sizes, widths, transformations, ...otherProps}) => {
-    const url = transformImageUrl(src, transformations)
+export const getImageProps = ({src, vwSizes, sizes: _sizes, srcSet: _srcSet, ...otherProps}) => {
+    const imageProps = {src: getSrcWithoutOptionalParams(src)}
 
-    const srcSet = (widths || [])
-        .map((width) => {
-            url.searchParams.set('sw', width)
-            return `${url} ${width}w`
-        })
-        .join(', ')
+    if (vwSizes) {
+        const sizes = mapVwSizesToSizes(vwSizes)
+        const srcSet = mapVwSizesToSrcSet(vwSizes)
+        imageProps.sizes = convertSizesToHTMLAttribute(sizes)
+        imageProps.srcSet = convertSrcSetToHTMLAttribute(srcSet, src)
+
+        if (vwSizes && _sizes && _srcSet) {
+            console.error('Cannot pass in sizes and srcSet when you already have specified vwSizes')
+        }
+    }
+
+    if (_sizes) {
+        imageProps.sizes = convertSizesToHTMLAttribute(_sizes)
+    }
+    if (_srcSet) {
+        imageProps.srcSet = convertSrcSetToHTMLAttribute(_srcSet, src)
+    }
 
     return {
-        src,
-        ...(sizes ? {sizes: buildSizes(sizes)} : {}),
-        ...(srcSet ? {srcSet} : {}),
+        ...imageProps,
         ...otherProps
     }
 }
 
 /**
- * @param {string} src - The image's url
- * @param {Object} [transformations] - How to transform the image's url
- * @return {URL} url object
+ * @param {(string[]|Object|string)} sizes
  */
-export const transformImageUrl = (src, transformations = {}) => {
-    const fallbackOrigin = getAppOrigin()
-    const url = new URL(src, fallbackOrigin)
-
-    for (const key in DEFAULT_TRANSFORMATIONS) {
-        url.searchParams.set(key, DEFAULT_TRANSFORMATIONS[key])
-    }
-    for (const key in transformations) {
-        url.searchParams.set(key, transformations[key])
+const convertSizesToHTMLAttribute = (sizes) => {
+    if (typeof sizes === 'string') {
+        return sizes
     }
 
-    return url
-}
+    // Convert to object first if needed
+    let _sizes = {}
+    if (Array.isArray(sizes)) {
+        sizes.forEach((size, i) => {
+            const key = breakpointLabels[i]
+            _sizes[key] = size
+        })
+    } else {
+        _sizes = sizes
+    }
 
-const DEFAULT_TRANSFORMATIONS = {
-    q: 60
-}
-
-const buildSizes = (sizes = {}) => {
+    // Then convert all sizes into string
+    const bp = theme.breakpoints
     const s = []
-
-    // TODO: can we get these programatically?
-    // Aligned with default Chakra breakpoints:
-    // https://chakra-ui.com/docs/features/responsive-styles
-    sizes['2xl'] && s.push(`(min-width: 96em) ${sizes['2xl']}`)
-    sizes.xl && s.push(`(min-width: 80em) ${sizes.xl}`)
-    sizes.lg && s.push(`(min-width: 62em) ${sizes.lg}`)
-    sizes.md && s.push(`(min-width: 48em) ${sizes.md}`)
-    sizes.sm && s.push(`(min-width: 30em) ${sizes.sm}`)
-    sizes.base && s.push(sizes.base)
+    _sizes['2xl'] && s.push(`(min-width: ${bp['2xl']}) ${_sizes['2xl']}`)
+    _sizes.xl && s.push(`(min-width: ${bp.xl}) ${_sizes.xl}`)
+    _sizes.lg && s.push(`(min-width: ${bp.lg}) ${_sizes.lg}`)
+    _sizes.md && s.push(`(min-width: ${bp.md}) ${_sizes.md}`)
+    _sizes.sm && s.push(`(min-width: ${bp.sm}) ${_sizes.sm}`)
+    _sizes.base && s.push(_sizes.base)
 
     return s.join(', ')
+}
+
+const breakpointLabels = ['base', 'sm', 'md', 'lg', 'xl', '2xl']
+
+/**
+ * @param {(number[]|string)} srcSet
+ * @param {string} srcFormat
+ */
+const convertSrcSetToHTMLAttribute = (srcSet, srcFormat) => {
+    if (typeof srcSet === 'string') {
+        return srcSet
+    }
+
+    const s = srcSet.map((imageWidth) => `${getSrc(srcFormat, imageWidth)} ${imageWidth}w`)
+    return s.join(', ')
+}
+
+/**
+ * @param {(number[]|Object)} vwSizes
+ */
+const mapVwSizesToSizes = (vwSizes) => {
+    return Array.isArray(vwSizes)
+        ? vwSizes.map((size) => `${size}vw`)
+        : responsivePropAsArray(vwSizes)
+}
+
+/**
+ * @param {(number[]|Object)} vwSizes
+ */
+const mapVwSizesToSrcSet = (vwSizes) => {
+    let sizes = Array.isArray(vwSizes) ? vwSizes.slice() : responsivePropAsArray(vwSizes)
+
+    if (sizes.length < breakpointLabels.length) {
+        const lastSize = sizes[sizes.length - 1]
+        const amountToPad = breakpointLabels.length - sizes.length
+
+        sizes = [...sizes, ...Array(amountToPad).fill(lastSize)]
+    }
+
+    const srcSet = []
+    sizes.forEach((size, i) => {
+        if (i === sizes.length - 1) {
+            return
+        }
+
+        const nextBp = breakpointLabels[i + 1]
+        const em = (size / 100) * parseFloat(theme.breakpoints[nextBp])
+        const px = emToPx(em)
+
+        srcSet.push(px)
+        srcSet.push(px * 2) // for devices with higher pixel density
+    })
+
+    return srcSet
+}
+
+/**
+ * @param {Object} prop
+ */
+const responsivePropAsArray = (prop) => {
+    let mostRecent
+    return breakpointLabels.map((bp) => {
+        if (prop[bp]) {
+            mostRecent = prop[bp]
+            return prop[bp]
+        } else {
+            return mostRecent
+        }
+    })
+}
+
+/**
+ * @param {number} em
+ * @param {number} browserDefaultFontSize
+ */
+const emToPx = (em, browserDefaultFontSize = 16) => Math.round(em * browserDefaultFontSize)
+
+/**
+ * @param {string} srcFormat
+ * @param {number} imageWidth
+ * @example
+ * // returns https://example.com/image_720.jpg
+ * getSrc('https://example.com/image[_{width}].jpg', 720)
+ */
+const getSrc = (srcFormat, imageWidth) => {
+    // 1. remove the surrounding []
+    // 2. replace {...} with imageWidth
+    return srcFormat.replace(/\[([^\]]+)\]/g, '$1').replace(/\{[^}]+\}/g, imageWidth)
+}
+
+/**
+ * @param {string} srcFormat
+ */
+const getSrcWithoutOptionalParams = (srcFormat) => {
+    const optionalParams = /\[[^\]]+\]/g
+    return srcFormat.replace(optionalParams, '')
 }
 
 // TODO: a simpler interface
